backend/requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-MySQL==1.5.2
numpy==1.21.5
uwsgi

